+++ src/main.rs
@@ -24,10 +24,10 @@ fn main() -> std::io::Result<()> {
 
     let image_aspect_ratio = 16.0 / 9.0;
     let render_image_width = 1024;
-    let render_image_height = (render_image_width as f32 / image_aspect_ratio) as i32;
+    let render_image_height = (render_image_width as f64 / image_aspect_ratio) as i32;
 
     let viewport_height = 2.0;
-    let image_ratio = render_image_width as f32 / render_image_height as f32;
+    let image_ratio = render_image_width as f64 / render_image_height as f64;
     let viewport_width = image_ratio * viewport_height; // use `image_ratio` for exact ratio
     let focal_length = 1.0;
 
@@ -42,8 +42,8 @@ fn main() -> std::io::Result<()> {
     let vp_origin_upper_left =
         camera_pos - (0.5 * vp_u_dir) - (0.5 * vp_v_dir) - Vec3::new(0.0, 0.0, focal_length);
 
-    let pixel_delta_u = vp_u_dir / render_image_width as f32;
-    let pixel_delta_v = vp_v_dir / render_image_height as f32;
+    let pixel_delta_u = vp_u_dir / render_image_width as f64;
+    let pixel_delta_v = vp_v_dir / render_image_height as f64;
     let pixel_00_loc = vp_origin_upper_left + 0.5 * (pixel_delta_u + pixel_delta_v);
 
     let mut world = World::new();
@@ -58,7 +58,7 @@ fn main() -> std::io::Result<()> {
         // eprintln!("Scan-lines processed: {}/{}", y, render_image_height);
 
         for x in 0..render_image_width {
-            let pixel_pos = pixel_00_loc + y as f32 * pixel_delta_v + x as f32 * pixel_delta_u;
+            let pixel_pos = pixel_00_loc + y as f64 * pixel_delta_v + x as f64 * pixel_delta_u;
 
             let ray = Ray::new(camera_pos, pixel_pos - camera_pos);
