+++ src/utils/math.rs
@@ -1,3 +1,3 @@
-pub fn degrees_ro_radians(deg: f32) -> f32 {
-    deg * std::f32::consts::PI / 180.0
+pub fn degrees_ro_radians(deg: f64) -> f64 {
+    deg * std::f64::consts::PI / 180.0
 }
